decouple test helper input struct from netlink library
netlink library has dependency on linux specific go libraries,
so in order to allow cross OS compilation moving input that will be
be used outside the agent (in ginkgo tests to send inputs) to a seperate
package.
